+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - SSH Hook: added support for the `com.hooks.ssh.pidfile_container` OCI annotation, which allows to customize the path to the Dropbear daemon PIDfile inside the container.
 - SSH Hook: added support for the `com.hooks.ssh.pidfile_host` OCI annotation, which optionally copies the PIDfile of the Dropbear server to the specified path on the host.
+- SSH Hook: added support for the `OVERLAY_MOUNT_HOME_SSH` environment variable, which allows to control the creation of an overlay filesystem on top of the container's `${HOME}/.ssh` directory.
+  More details [here](https://sarus.readthedocs.io/en/stable/config/ssh-hook.html#hook-configuration)
 
 ## [1.6.1]